Fixed integer bug

+++ CSLua/CS-Lua/Drawing.h
@@ -8,7 +8,7 @@ class CDrawing
 {
 public:
 
-	void DrawString(unsigned long font, int x, int y,  const char* text)
+	void DrawString(unsigned long font, float x, float y,  const char* text)
 	{
 		std::wstring_convert<std::codecvt_utf8_utf16<wchar_t>> converter;
 		std::wstring wtext = converter.from_bytes(text);
@@ -18,7 +18,7 @@ public:
 		g_pSurface->DrawText(wtext.c_str(), wtext.size());
 	}
 
-	unsigned long CreateFont(const char* fontname, int size, bool bold, bool italic, bool outlined, bool dropshadow)
+	unsigned long CreateFont(const char* fontname, float size, bool bold, bool italic, bool outlined, bool dropshadow)
 	{
 		unsigned long font = g_pSurface->CreateFont();
 		int flags = (italic ? FONTFLAG_ITALIC : 0) | (outlined ? FONTFLAG_OUTLINE : 0) | (dropshadow ? FONTFLAG_DROPSHADOW : 0) | FONTFLAG_ANTIALIAS;
@@ -26,17 +26,17 @@ public:
 		return font;
 	}
 
-	void DrawLine(int x, int y, int x2, int y2)
+	void DrawLine(float x, float y, float x2, float y2)
 	{
 		g_pSurface->DrawLine(x, y, x2, y2);
 	}
 
-	void DrawFilledRect(int x, int y, int w, int h)
+	void DrawFilledRect(float x, float y, float w, float h)
 	{
 		g_pSurface->DrawFilledRect(x, y, x + w, y + h);
 	}
 
-	void DrawOutlinedRect(int x, int y, int w, int h, int r, int g, int b, int a)
+	void DrawOutlinedRect(float x, float y, float w, float h, float r, float g, float b, float a)
 	{
 		g_pSurface->DrawOutlinedRect(x, y, x + w, y + h);
 		g_pSurface->DrawSetcolor(r, g, b, a);
@@ -49,12 +49,12 @@ public:
 		g_pSurface->DrawSetcolor(m_r, m_g, m_b, m_a);
 	}
 
-	void DrawRect(int x, int y, int w, int h)
+	void DrawRect(float x, float y, float w, float h)
 	{
 		g_pSurface->DrawOutlinedRect(x, y, x + w, y + h);
 	}
 
-	void SetDrawColor(int r, int g, int b, int a)
+	void SetDrawColor(float r, float g, float b, float a)
 	{
 		m_r = r; m_g = g; m_b = b; m_a = a;
 		g_pSurface->DrawSetcolor(r, g, b, a);
